+++ hydra-node/src/Hydra/Chain/Blockfrost/Client.hs
@@ -1,19 +1,42 @@
 {-# LANGUAGE RecordWildCards #-}
 
-module Hydra.Chain.Blockfrost.Client where
+module Hydra.Chain.Blockfrost.Client (
+  module Hydra.Chain.Blockfrost.Client,
+  module Blockfrost.Client,
+) where
 
 import Hydra.Prelude
 
 import Blockfrost.Client (
+  Block (..),
+  BlockHash (..),
   BlockfrostClientT,
+  Genesis (..),
+  Project,
+  Slot (..),
+  TransactionCBOR (..),
+  TxHashCBOR (..),
+  allPages,
+  def,
+  getBlock,
+  getBlockTxsCBOR',
+  getLedgerGenesis,
+  listPools,
+  projectFromFile,
   runBlockfrost,
+  tryError,
+  unBlockHash,
+  unSlot,
  )
 import Blockfrost.Client qualified as Blockfrost
+import Cardano.Chain.Genesis (mainnetProtocolMagicId)
+import Cardano.Crypto.ProtocolMagic (ProtocolMagicId (..))
 import Data.Map.Strict qualified as Map
 import Data.Time.Clock.POSIX
-import Hydra.Cardano.Api hiding (LedgerState, fromNetworkMagic)
+import Hydra.Cardano.Api hiding (LedgerState, fromNetworkMagic, queryGenesisParameters)
 
 import Cardano.Api.UTxO qualified as UTxO
+import Cardano.Crypto.Hash (hashToTextAsHex)
 import Cardano.Ledger.Api.PParams
 import Cardano.Ledger.BaseTypes (EpochInterval (..), EpochSize (..), NonNegativeInterval, UnitInterval, boundRational, unsafeNonZero)
 import Cardano.Ledger.Binary.Version (mkVersion)
@@ -35,22 +58,31 @@ import Cardano.Ledger.Plutus.CostModels (CostModels, mkCostModel, mkCostModels)
 import Cardano.Ledger.Shelley.API (ProtVer (..))
 import Cardano.Slotting.Time (RelativeTime (..), mkSlotLength)
 import Control.Lens ((.~), (^.))
-import Data.Default (def)
+import Data.List qualified as List
 import Data.SOP.NonEmpty (nonEmptyFromList)
 import Data.Set qualified as Set
 import Data.Text qualified as T
 import Hydra.Cardano.Api.Prelude (StakePoolKey, fromNetworkMagic)
-import Hydra.Chain.CardanoClient (QueryPoint (..))
-import Hydra.Chain.ScriptRegistry (buildScriptPublishingTxs)
-import Hydra.Tx (txId)
+import Hydra.Tx (ScriptRegistry, newScriptRegistry)
 import Money qualified
 import Ouroboros.Consensus.Block (GenesisWindow (..))
 import Ouroboros.Consensus.HardFork.History (Bound (..), EraEnd (..), EraParams (..), EraSummary (..), SafeZone (..), Summary (..), mkInterpreter)
 
-data APIBlockfrostError
-  = BlockfrostError Text
+data BlockfrostException
+  = TimeoutOnUTxO TxId
+  | FailedToDecodeAddress Text
+  | ByronAddressNotSupported
+  | FailedUTxOForHash Text
+  | FailedEraHistory
+  | AssetNameMissing
+  | DeserialiseError Text
   | DecodeError Text
+  | BlockfrostAPIError Text
   deriving (Show, Exception)
+
+newtype APIBlockfrostError
+  = BlockfrostError BlockfrostException
+  deriving newtype (Show, Exception)
 
 runBlockfrostM ::
   (MonadIO m, MonadThrow m) =>
@@ -60,156 +92,38 @@ runBlockfrostM ::
 runBlockfrostM prj action = do
   result <- liftIO $ runBlockfrost prj action
   case result of
-    Left err -> throwIO (BlockfrostError $ show err)
+    Left err -> throwIO $ BlockfrostError $ BlockfrostAPIError (show err)
     Right val -> pure val
 
-publishHydraScripts ::
-  -- | The path where the Blockfrost project token hash is stored.
-  FilePath ->
-  -- | Keys assumed to hold funds to pay for the publishing transaction.
-  SigningKey PaymentKey ->
-  IO [TxId]
-publishHydraScripts projectPath sk = do
-  prj <- Blockfrost.projectFromFile projectPath
-  runBlockfrostM prj $ do
-    Blockfrost.Genesis
-      { _genesisNetworkMagic = networkMagic
-      , _genesisSystemStart = systemStart'
-      } <-
-      queryGenesis
-    pparams <- toCardanoPParams
-    let address = Blockfrost.Address (vkAddress networkMagic)
-    let networkId = toCardanoNetworkId networkMagic
-    let changeAddress = mkVkAddress networkId vk
-    stakePools' <- Blockfrost.listPools
-    let stakePools = Set.fromList (toCardanoPoolId <$> stakePools')
-    let systemStart = SystemStart $ posixSecondsToUTCTime systemStart'
-    eraHistory <- mkEraHistory
-    utxo <- Blockfrost.getAddressUtxos address
-    let cardanoUTxO = toCardanoUTxO utxo changeAddress
-
-    txs <- liftIO $ buildScriptPublishingTxs pparams systemStart networkId eraHistory stakePools cardanoUTxO sk
-    forM txs $ \(tx :: Tx) -> do
-      void $ Blockfrost.submitTx $ Blockfrost.CBORString $ fromStrict $ serialiseToCBOR tx
-      pure $ txId tx
+-- | Query for 'TxIn's in the search for outputs containing all the reference
+-- scripts of the 'ScriptRegistry'.
+--
+-- This is implemented by repeated querying until we have all necessary
+-- reference scripts as we do only know the transaction id, not the indices.
+--
+-- Can throw at least 'NewScriptRegistryException' on failure.
+queryScriptRegistry ::
+  [TxId] ->
+  BlockfrostClientT IO ScriptRegistry
+queryScriptRegistry txIds = do
+  Blockfrost.Genesis
+    { _genesisNetworkMagic
+    , _genesisSystemStart
+    } <-
+    queryGenesisParameters
+  let networkId = toCardanoNetworkId _genesisNetworkMagic
+  utxoList <- forM candidates $ \(TxIn (TxId candidateHash) _) -> queryUTxOByTxIn networkId $ hashToTextAsHex candidateHash
+  case newScriptRegistry $ fold utxoList of
+    Left e -> liftIO $ throwIO e
+    Right sr -> pure sr
  where
-  vk = getVerificationKey sk
+  candidates = map (\txid -> TxIn txid (TxIx 0)) txIds
 
-  vkAddress networkMagic = textAddrOf (toCardanoNetworkId networkMagic) vk
-
-scriptTypeToPlutusVersion :: Blockfrost.ScriptType -> Maybe Language
-scriptTypeToPlutusVersion = \case
-  Blockfrost.PlutusV1 -> Just PlutusV1
-  Blockfrost.PlutusV2 -> Just PlutusV2
-  Blockfrost.PlutusV3 -> Just PlutusV3
-  Blockfrost.Timelock -> Nothing
-
-toCardanoPoolId :: Blockfrost.PoolId -> Hash StakePoolKey
-toCardanoPoolId (Blockfrost.PoolId textPoolId) =
-  case deserialiseFromRawBytesHex (encodeUtf8 textPoolId) of
-    Left err -> error (show err)
-    Right pool -> pool
-
-toCardanoUTxO :: [Blockfrost.AddressUtxo] -> AddressInEra -> UTxO' (TxOut CtxUTxO)
-toCardanoUTxO utxos addr = UTxO.fromList (toEntry <$> utxos)
- where
-  toEntry :: Blockfrost.AddressUtxo -> (TxIn, TxOut CtxUTxO)
-  toEntry utxo = (toCardanoTxIn utxo, toCardanoTxOut utxo addr)
-
-toCardanoTxIn :: Blockfrost.AddressUtxo -> TxIn
-toCardanoTxIn Blockfrost.AddressUtxo{_addressUtxoTxHash = Blockfrost.TxHash{unTxHash}, _addressUtxoOutputIndex} =
-  case deserialiseFromRawBytesHex (encodeUtf8 unTxHash) of
-    Left err -> error (show err)
-    Right txid -> TxIn txid (TxIx (fromIntegral _addressUtxoOutputIndex))
-
--- REVIEW! TxOutDatumNone and ReferenceScriptNone
-toCardanoTxOut :: Blockfrost.AddressUtxo -> AddressInEra -> TxOut CtxUTxO
-toCardanoTxOut addrUTxO addr =
-  TxOut addr (toCardanoValue _addressUtxoAmount) TxOutDatumNone ReferenceScriptNone
- where
-  Blockfrost.AddressUtxo{_addressUtxoAmount, _addressUtxoDataHash, _addressUtxoInlineDatum, _addressUtxoReferenceScriptHash} = addrUTxO
-
-toCardanoPolicyId :: Text -> PolicyId
-toCardanoPolicyId pid =
-  case deserialiseFromRawBytesHex (encodeUtf8 pid) of
-    Left err -> error (show err)
-    Right p -> p
-
-toCardanoAssetName :: Text -> AssetName
-toCardanoAssetName = AssetName . encodeUtf8
-
-toCardanoValue :: [Blockfrost.Amount] -> Value
-toCardanoValue = foldMap convertAmount
- where
-  convertAmount (Blockfrost.AdaAmount lovelaces) =
-    fromList
-      [
-        ( AdaAssetId
-        , Quantity (toInteger lovelaces)
-        )
-      ]
-  convertAmount (Blockfrost.AssetAmount money) =
-    let currency = Money.someDiscreteCurrency money
-     in fromList
-          [
-            ( AssetId
-                (toCardanoPolicyId currency)
-                (toCardanoAssetName currency)
-            , Quantity (Money.someDiscreteAmount money)
-            )
-          ]
-
--- ** Helpers
-
-unwrapAddress :: AddressInEra -> Text
-unwrapAddress = \case
-  ShelleyAddressInEra addr -> serialiseToBech32 addr
-  ByronAddressInEra{} -> error "Byron."
-
-textAddrOf :: NetworkId -> VerificationKey PaymentKey -> Text
-textAddrOf networkId vk = unwrapAddress (mkVkAddress @Era networkId vk)
-
-toCardanoNetworkId :: Integer -> NetworkId
-toCardanoNetworkId = \case
-  0 -> Mainnet
-  magicNbr -> Testnet (NetworkMagic (fromInteger magicNbr))
-
-data BlockfrostConversion
-  = BlockfrostConversion
-  { a0 :: NonNegativeInterval
-  , rho :: UnitInterval
-  , tau :: UnitInterval
-  , priceMemory :: NonNegativeInterval
-  , priceSteps :: NonNegativeInterval
-  , pvtMotionNoConfidence :: UnitInterval
-  , pvtCommitteeNormal :: UnitInterval
-  , pvtCommitteeNoConfidence :: UnitInterval
-  , pvtHardForkInitiation :: UnitInterval
-  , pvtPPSecurityGroup :: UnitInterval
-  , dvtMotionNoConfidence :: UnitInterval
-  , dvtCommitteeNormal :: UnitInterval
-  , dvtCommitteeNoConfidence :: UnitInterval
-  , dvtUpdateToConstitution :: UnitInterval
-  , dvtHardForkInitiation :: UnitInterval
-  , dvtPPNetworkGroup :: UnitInterval
-  , dvtPPEconomicGroup :: UnitInterval
-  , dvtPPTechnicalGroup :: UnitInterval
-  , dvtPPGovGroup :: UnitInterval
-  , dvtTreasuryWithdrawal :: UnitInterval
-  , committeeMinSize :: Blockfrost.Quantity
-  , committeeMaxTermLength :: Blockfrost.Quantity
-  , govActionLifetime :: Blockfrost.Quantity
-  , govActionDeposit :: Coin
-  , drepDeposit :: Integer
-  , drepActivity :: Blockfrost.Quantity
-  , minFeeRefScriptCostPerByte :: NonNegativeInterval
-  }
-
-toCardanoPParams :: MonadIO m => BlockfrostClientT m (PParams LedgerEra)
-toCardanoPParams = do
+queryProtocolParameters :: MonadIO m => BlockfrostClientT m (PParams LedgerEra)
+queryProtocolParameters = do
   pparams <- Blockfrost.getLatestEpochProtocolParams
-  minVersion <- liftIO $ mkVersion $ pparams ^. Blockfrost.protocolMinorVer
-  let maxVersion = fromIntegral $ pparams ^. Blockfrost.protocolMajorVer
+  let minVersion = fromIntegral $ pparams ^. Blockfrost.protocolMinorVer
+  maxVersion <- liftIO $ mkVersion $ pparams ^. Blockfrost.protocolMajorVer
   let results = do
         a0 <- boundRational (pparams ^. Blockfrost.a0)
         rho <- boundRational (pparams ^. Blockfrost.rho)
@@ -257,13 +171,13 @@ toCardanoPParams = do
           & ppA0L .~ a0
           & ppRhoL .~ rho
           & ppTauL .~ tau
-          & ppProtocolVersionL .~ ProtVer minVersion maxVersion
+          & ppProtocolVersionL .~ ProtVer maxVersion minVersion
           & ppMinPoolCostL .~ fromIntegral (pparams ^. Blockfrost.minPoolCost)
           & ppCoinsPerUTxOByteL .~ CoinPerByte (fromIntegral (pparams ^. Blockfrost.coinsPerUtxoSize))
-          & ppCostModelsL .~ convertCostModels (pparams ^. Blockfrost.costModels)
+          & ppCostModelsL .~ convertCostModels (pparams ^. Blockfrost.costModelsRaw)
           & ppPricesL .~ Prices priceMemory priceSteps
-          & ppMaxTxExUnitsL .~ ExUnits (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxTxExSteps) (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxTxExMem)
-          & ppMaxBlockExUnitsL .~ ExUnits (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxBlockExSteps) (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxBlockExMem)
+          & ppMaxTxExUnitsL .~ ExUnits (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxTxExMem) (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxTxExSteps)
+          & ppMaxBlockExUnitsL .~ ExUnits (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxBlockExMem) (fromIntegral $ Blockfrost.unQuantity $ pparams ^. Blockfrost.maxBlockExSteps)
           & ppMaxValSizeL .~ fromIntegral (Blockfrost.unQuantity $ pparams ^. Blockfrost.maxValSize)
           & ppCollateralPercentageL .~ fromIntegral (pparams ^. Blockfrost.collateralPercent)
           & ppMaxCollateralInputsL .~ fromIntegral (pparams ^. Blockfrost.maxCollateralInputs)
@@ -277,21 +191,156 @@ toCardanoPParams = do
           & ppDRepActivityL .~ EpochInterval (fromIntegral $ Blockfrost.unQuantity drepActivity)
           & ppMinFeeRefScriptCostPerByteL .~ minFeeRefScriptCostPerByte
  where
-  convertCostModels :: Blockfrost.CostModels -> CostModels
+  convertCostModels :: Blockfrost.CostModelsRaw -> CostModels
   convertCostModels costModels =
-    let costModelsMap = Blockfrost.unCostModels costModels
+    let costModelsMap = Blockfrost.unCostModelsRaw costModels
      in foldMap
           ( (mempty <>)
               . ( \(scriptType, v) ->
                     case scriptTypeToPlutusVersion scriptType of
                       Nothing -> mempty
                       Just plutusScript ->
-                        case mkCostModel plutusScript (fromIntegral <$> Map.elems v) of
+                        case mkCostModel plutusScript (fromIntegral <$> v) of
                           Left _ -> mempty
                           Right costModel -> mkCostModels $ Map.singleton plutusScript costModel
                 )
           )
           (Map.toList costModelsMap)
+
+-- ** Helpers
+
+toCardanoUTxO :: NetworkId -> TxIn -> Blockfrost.Address -> Maybe Blockfrost.ScriptHash -> Maybe Blockfrost.DatumHash -> [Blockfrost.Amount] -> Maybe Blockfrost.InlineDatum -> BlockfrostClientT IO (UTxO' (TxOut ctx))
+toCardanoUTxO networkId txIn address scriptHash datumHash amount inlineDatum = do
+  let addrTxt = Blockfrost.unAddress address
+  let datumHash' = Blockfrost.unDatumHash <$> datumHash
+  let inlineDatum' = Blockfrost._scriptDatumCborCbor . Blockfrost.unInlineDatum <$> inlineDatum
+  val <- toCardanoValue amount
+  plutusScript <- maybe (pure Nothing) (queryScript . Blockfrost.unScriptHash) scriptHash
+  o <- toCardanoTxOut networkId addrTxt val datumHash' inlineDatum' plutusScript
+  pure $ UTxO.singleton txIn o
+
+scriptTypeToPlutusVersion :: Blockfrost.ScriptType -> Maybe Language
+scriptTypeToPlutusVersion = \case
+  Blockfrost.PlutusV1 -> Just PlutusV1
+  Blockfrost.PlutusV2 -> Just PlutusV2
+  Blockfrost.PlutusV3 -> Just PlutusV3
+  Blockfrost.Timelock -> Nothing
+
+toCardanoPoolId :: Blockfrost.PoolId -> Hash StakePoolKey
+toCardanoPoolId (Blockfrost.PoolId textPoolId) =
+  case deserialiseFromRawBytesHex (encodeUtf8 textPoolId) of
+    Left err -> error (show err)
+    Right pool -> pool
+
+toCardanoTxIn :: Text -> Integer -> TxIn
+toCardanoTxIn txHash i =
+  case deserialiseFromRawBytesHex (encodeUtf8 txHash) of
+    Left err -> error (show err)
+    Right txid -> TxIn txid (TxIx (fromIntegral i))
+
+toCardanoTxOut :: NetworkId -> Text -> Value -> Maybe Text -> Maybe Text -> Maybe PlutusScript -> BlockfrostClientT IO (TxOut ctx)
+toCardanoTxOut networkId addrTxt val mDatumHash mInlineDatum plutusScript = do
+  let datum =
+        case mInlineDatum of
+          Nothing ->
+            case mDatumHash of
+              Nothing -> TxOutDatumNone
+              Just datumHash -> TxOutDatumHash (fromString $ T.unpack datumHash)
+          Just cborDatum ->
+            case deserialiseFromCBOR (proxyToAsType (Proxy @HashableScriptData)) (encodeUtf8 cborDatum) of
+              Left _ -> TxOutDatumNone
+              Right hashableScriptData -> TxOutDatumInline hashableScriptData
+  case plutusScript of
+    Nothing -> do
+      case toCardanoAddress addrTxt of
+        Nothing -> liftIO $ throwIO $ BlockfrostError $ FailedToDecodeAddress addrTxt
+        Just addr -> pure $ TxOut addr val datum ReferenceScriptNone
+    Just script -> pure $ TxOut (scriptAddr script) val datum (mkScriptRef script)
+ where
+  scriptAddr script =
+    makeShelleyAddressInEra
+      shelleyBasedEra
+      networkId
+      (PaymentCredentialByScript $ hashScript $ PlutusScript script)
+      NoStakeAddress
+
+toCardanoPolicyIdAndAssetName :: Text -> BlockfrostClientT IO (PolicyId, AssetName)
+toCardanoPolicyIdAndAssetName pid = do
+  Blockfrost.AssetDetails{_assetDetailsPolicyId, _assetDetailsAssetName} <- Blockfrost.getAssetDetails (Blockfrost.mkAssetId pid)
+  case deserialiseFromRawBytesHex (encodeUtf8 $ Blockfrost.unPolicyId _assetDetailsPolicyId) of
+    Left err -> liftIO $ throwIO $ BlockfrostError $ DeserialiseError (show err)
+    Right p ->
+      case _assetDetailsAssetName of
+        Nothing -> liftIO $ throwIO $ BlockfrostError AssetNameMissing
+        Just assetName ->
+          case deserialiseFromRawBytesHex (encodeUtf8 assetName) of
+            Left err -> liftIO $ throwIO $ BlockfrostError $ DeserialiseError (show err)
+            Right asset -> pure (p, asset)
+
+toCardanoValue :: [Blockfrost.Amount] -> BlockfrostClientT IO Value
+toCardanoValue = foldMapM convertAmount
+ where
+  convertAmount (Blockfrost.AdaAmount lovelaces) =
+    pure $
+      fromList
+        [
+          ( AdaAssetId
+          , Quantity (toInteger lovelaces)
+          )
+        ]
+  convertAmount (Blockfrost.AssetAmount money) = do
+    let currency = Money.someDiscreteCurrency money
+    (cardanoPolicyId, assetName) <- toCardanoPolicyIdAndAssetName currency
+    pure $
+      fromList
+        [
+          ( AssetId
+              cardanoPolicyId
+              assetName
+          , Quantity (Money.someDiscreteAmount money)
+          )
+        ]
+
+toCardanoAddress :: Text -> Maybe AddressInEra
+toCardanoAddress addrTxt =
+  ShelleyAddressInEra <$> deserialiseAddress (AsAddress AsShelleyAddr) addrTxt
+
+toCardanoNetworkId :: Integer -> NetworkId
+toCardanoNetworkId magic =
+  if fromIntegral magic == unProtocolMagicId mainnetProtocolMagicId
+    then Mainnet
+    else Testnet (NetworkMagic (fromInteger magic))
+
+data BlockfrostConversion
+  = BlockfrostConversion
+  { a0 :: NonNegativeInterval
+  , rho :: UnitInterval
+  , tau :: UnitInterval
+  , priceMemory :: NonNegativeInterval
+  , priceSteps :: NonNegativeInterval
+  , pvtMotionNoConfidence :: UnitInterval
+  , pvtCommitteeNormal :: UnitInterval
+  , pvtCommitteeNoConfidence :: UnitInterval
+  , pvtHardForkInitiation :: UnitInterval
+  , pvtPPSecurityGroup :: UnitInterval
+  , dvtMotionNoConfidence :: UnitInterval
+  , dvtCommitteeNormal :: UnitInterval
+  , dvtCommitteeNoConfidence :: UnitInterval
+  , dvtUpdateToConstitution :: UnitInterval
+  , dvtHardForkInitiation :: UnitInterval
+  , dvtPPNetworkGroup :: UnitInterval
+  , dvtPPEconomicGroup :: UnitInterval
+  , dvtPPTechnicalGroup :: UnitInterval
+  , dvtPPGovGroup :: UnitInterval
+  , dvtTreasuryWithdrawal :: UnitInterval
+  , committeeMinSize :: Blockfrost.Quantity
+  , committeeMaxTermLength :: Blockfrost.Quantity
+  , govActionLifetime :: Blockfrost.Quantity
+  , govActionDeposit :: Coin
+  , drepDeposit :: Integer
+  , drepActivity :: Blockfrost.Quantity
+  , minFeeRefScriptCostPerByte :: NonNegativeInterval
+  }
 
 toCardanoGenesisParameters :: Blockfrost.Genesis -> GenesisParameters ShelleyEra
 toCardanoGenesisParameters bfGenesis =
@@ -322,14 +371,21 @@ toCardanoGenesisParameters bfGenesis =
     , _genesisSecurityParam
     } = bfGenesis
 
-mkEraHistory :: BlockfrostClientT IO EraHistory
-mkEraHistory = do
+submitTransaction :: MonadIO m => Tx -> BlockfrostClientT m Blockfrost.TxHash
+submitTransaction tx = Blockfrost.submitTx $ Blockfrost.CBORString $ fromStrict $ serialiseToCBOR tx
+
+----------------
+-- Queries --
+----------------
+
+queryEraHistory :: BlockfrostClientT IO EraHistory
+queryEraHistory = do
   eras' <- Blockfrost.getNetworkEras
-  let eras = filter withoutEmptyEra eras'
+  let eras = filter isEmptyEra eras'
   let summary = mkEra <$> eras
   case nonEmptyFromList summary of
     Nothing ->
-      liftIO $ throwIO $ BlockfrostError "Failed to create EraHistory."
+      liftIO $ throwIO $ BlockfrostError FailedEraHistory
     Just s -> pure $ EraHistory (mkInterpreter $ Summary s)
  where
   mkBound Blockfrost.NetworkEraBound{_boundEpoch, _boundSlot, _boundTime} =
@@ -351,45 +407,98 @@ mkEraHistory = do
       , eraEnd = EraEnd $ mkBound _networkEraEnd
       , eraParams = mkEraParams _networkEraParameters
       }
-  withoutEmptyEra
+  isEmptyEra
     Blockfrost.NetworkEraSummary
       { _networkEraStart = Blockfrost.NetworkEraBound{_boundTime = boundStart}
       , _networkEraEnd = Blockfrost.NetworkEraBound{_boundTime = boundEnd}
-      } = boundStart == 0 && boundEnd == 0
+      , _networkEraParameters
+      } = boundStart /= 0 && boundEnd /= 0
 
-----------------
--- Wallet API --
-----------------
+-- | Query the Blockfrost API to get the 'UTxO' for 'TxIn' and convert to cardano 'UTxO'.
+queryUTxOByTxIn :: NetworkId -> Text -> BlockfrostClientT IO UTxO
+queryUTxOByTxIn networkId txHash = go (300 :: Int) -- TODO: make this configurable
+ where
+  go 0 = liftIO $ throwIO $ BlockfrostError $ FailedUTxOForHash txHash
+  go n = do
+    res <- Blockfrost.tryError $ Blockfrost.getTxUtxos (Blockfrost.TxHash txHash)
+    case res of
+      Left _e -> liftIO (threadDelay 1) >> go (n - 1)
+      Right Blockfrost.TransactionUtxos{_transactionUtxosInputs, _transactionUtxosOutputs} ->
+        foldMapM
+          ( \Blockfrost.UtxoOutput{_utxoOutputOutputIndex, _utxoOutputAddress, _utxoOutputAmount, _utxoOutputDataHash, _utxoOutputInlineDatum, _utxoOutputReferenceScriptHash} ->
+              let txIn = toCardanoTxIn txHash _utxoOutputOutputIndex
+               in toCardanoUTxO networkId txIn _utxoOutputAddress _utxoOutputReferenceScriptHash _utxoOutputDataHash _utxoOutputAmount _utxoOutputInlineDatum
+          )
+          _transactionUtxosOutputs
+
+queryScript :: Text -> BlockfrostClientT IO (Maybe PlutusScript)
+queryScript scriptHashTxt = do
+  Blockfrost.ScriptCBOR{_scriptCborCbor} <- Blockfrost.getScriptCBOR $ Blockfrost.ScriptHash scriptHashTxt
+  case _scriptCborCbor of
+    Nothing -> pure Nothing
+    Just fullScriptCBOR ->
+      case decodeBase16 fullScriptCBOR :: Either String ByteString of
+        Left _ -> pure Nothing
+        Right bytes ->
+          case deserialiseFromCBOR (proxyToAsType (Proxy @PlutusScript)) bytes of
+            Left _ -> pure Nothing
+            Right plutusScript -> pure $ Just plutusScript
 
 -- | Query the Blockfrost API for address UTxO and convert to cardano 'UTxO'.
-queryUTxO :: SigningKey PaymentKey -> NetworkId -> BlockfrostClientT IO UTxO
-queryUTxO sk networkId = do
-  let address = Blockfrost.Address vkAddress
-  utxo <- Blockfrost.getAddressUtxos address
-  let cardanoAddress = mkVkAddress networkId vk
-  pure $ toCardanoUTxO utxo cardanoAddress
- where
-  vk = getVerificationKey sk
-  vkAddress = textAddrOf networkId vk
+-- NOTE: We accept the address list here to be compatible with cardano-api but in
+-- fact this is a single address query always.
+queryUTxO :: NetworkId -> [Address ShelleyAddr] -> BlockfrostClientT IO UTxO
+queryUTxO networkId addresses = do
+  let address' = Blockfrost.Address . serialiseAddress $ List.head addresses
+  utxoWithAddresses <- Blockfrost.getAddressUtxos' address' (Blockfrost.paged 1 1) Blockfrost.desc
+
+  foldMapM
+    ( \Blockfrost.AddressUtxo
+         { Blockfrost._addressUtxoAddress
+         , Blockfrost._addressUtxoTxHash = Blockfrost.TxHash{unTxHash}
+         , Blockfrost._addressUtxoOutputIndex
+         , Blockfrost._addressUtxoAmount
+         , Blockfrost._addressUtxoBlock
+         , Blockfrost._addressUtxoDataHash
+         , Blockfrost._addressUtxoInlineDatum
+         , Blockfrost._addressUtxoReferenceScriptHash
+         } ->
+          let txin = toCardanoTxIn unTxHash _addressUtxoOutputIndex
+           in toCardanoUTxO networkId txin _addressUtxoAddress _addressUtxoReferenceScriptHash _addressUtxoDataHash _addressUtxoAmount _addressUtxoInlineDatum
+    )
+    utxoWithAddresses
+
+queryUTxOFor :: VerificationKey PaymentKey -> BlockfrostClientT IO UTxO
+queryUTxOFor vk = do
+  Blockfrost.Genesis
+    { _genesisNetworkMagic = networkMagic
+    } <-
+    queryGenesisParameters
+  let networkId = toCardanoNetworkId networkMagic
+  case mkVkAddress networkId vk of
+    ShelleyAddressInEra addr ->
+      queryUTxO networkId [addr]
+    ByronAddressInEra{} ->
+      liftIO $ throwIO $ BlockfrostError ByronAddressNotSupported
 
 -- | Query the Blockfrost API for 'Genesis'
-queryGenesis :: BlockfrostClientT IO Blockfrost.Genesis
-queryGenesis = Blockfrost.getLedgerGenesis
+queryGenesisParameters :: BlockfrostClientT IO Blockfrost.Genesis
+queryGenesisParameters = Blockfrost.getLedgerGenesis
+
+querySystemStart :: BlockfrostClientT IO SystemStart
+querySystemStart = do
+  Blockfrost.Genesis{_genesisSystemStart} <- queryGenesisParameters
+  pure $ SystemStart $ posixSecondsToUTCTime _genesisSystemStart
 
 -- | Query the Blockfrost API for 'Genesis' and convert to cardano 'ChainPoint'.
-queryTip :: QueryPoint -> BlockfrostClientT IO ChainPoint
-queryTip queryPoint = do
+queryTip :: BlockfrostClientT IO ChainPoint
+queryTip = do
   Blockfrost.Block
     { _blockHeight
     , _blockHash
     , _blockSlot
-    } <- case queryPoint of
-    QueryTip -> Blockfrost.getLatestBlock
-    QueryAt point -> do
-      let slot = case point of
-            ChainPointAtGenesis -> 0
-            ChainPoint slotNo _ -> fromIntegral $ unSlotNo slotNo
-      Blockfrost.getBlock (Left slot)
+    } <-
+    Blockfrost.getLatestBlock
   let slotAndBlockNumber = do
         blockSlot <- _blockSlot
         blockNumber <- _blockHeight
@@ -404,3 +513,41 @@ queryTip queryPoint = do
             (SlotNo $ fromIntegral $ Blockfrost.unSlot blockSlot)
             (fromString $ T.unpack blockHash)
             (BlockNo $ fromIntegral blockNo)
+
+queryStakePools ::
+  BlockfrostClientT IO (Set PoolId)
+queryStakePools = do
+  stakePools' <- Blockfrost.listPools
+  pure $ Set.fromList (toCardanoPoolId <$> stakePools')
+
+awaitTransaction :: Tx -> BlockfrostClientT IO UTxO
+awaitTransaction tx = do
+  Blockfrost.Genesis{_genesisNetworkMagic} <- queryGenesisParameters
+  let networkId = toCardanoNetworkId _genesisNetworkMagic
+  let TxId txhash = getTxId $ getTxBody tx
+  queryUTxOByTxIn networkId (hashToTextAsHex txhash)
+
+-- | Await for specific UTxO at address - the one that is produced by the given 'TxId'.
+awaitUTxO ::
+  -- | Network id
+  NetworkId ->
+  -- | Address we are interested in
+  [Address ShelleyAddr] ->
+  -- | Last transaction ID to await
+  TxId ->
+  -- | Number of seconds to wait
+  Int ->
+  BlockfrostClientT IO UTxO
+awaitUTxO networkId addresses txid i = do
+  go i
+ where
+  go 0 = liftIO $ throwIO $ BlockfrostError (TimeoutOnUTxO txid)
+  go n = do
+    utxo <- Blockfrost.tryError $ queryUTxO networkId addresses
+    case utxo of
+      Left _e -> liftIO (threadDelay 1) >> go (n - 1)
+      Right utxo' ->
+        let wantedUTxO = UTxO.fromList $ List.filter (\(TxIn txid' _, _) -> txid' == txid) (UTxO.toList utxo')
+         in if null wantedUTxO
+              then liftIO (threadDelay 1) >> go (n - 1)
+              else pure utxo'